+++ pytsite/auth/__init__.py
@@ -47,9 +47,6 @@ __init()
 
 
 # Public API
-from . import _error, _functions, _model
-error = _error
-model = _model
-
+from . import _error as error, _functions, _model as model
 from ._functions import define_permission_group,  define_permission, get_current_user, get_permission, \
-    get_permission_groups, get_permissions, get_user_statuses, get_permission_group
+    get_permission_groups, get_permissions, get_user_statuses, get_permission_group, get_user, create_user, get_role