+++ pytsite/route_alias/_manager.py
@@ -9,7 +9,7 @@ from pytsite.core import util as _util, lang as _lang, odm as _odm
 from . import _model
 
 
-def create(alias: str, target: str=None) -> _model.RouteAliasModel:
+def create(alias: str, target: str=None) -> _model.RouteAlias:
     """Create a route alias instance.
     """
     entity = _odm.dispense('route_alias')
@@ -41,7 +41,7 @@ def sanitize_alias_string(string: str) -> str:
         else:
             string = re.sub(r'-\d+$', '-' + str(itr), string)
 
-def find_by_target(target: str) -> _model.RouteAliasModel:
+def find_by_target(target: str) -> _model.RouteAlias:
     """Find route alias by target.
     """
     lng = _lang.get_current_lang()